Refactor: テーマ入力の部分をコンポーネント化
